added States entity without foreign key

+++ src/main/kotlin/com/itmo/mpa/entity/Diseases.java
@@ -9,7 +9,7 @@ public class Diseases {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "DiseaseId", nullable = false, updatable = false)
-    private int diseaseId;
+    private long diseaseId;
 
     @Column(name = "Name", nullable = false)
     private String name;
